Fix the NSBrowser delegate methods in NSSavePanel to have the correct types.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36061 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSSavePanel.m
@@ -552,7 +552,7 @@ setPath(NSBrowser *browser, NSString *path)
 
 - (BOOL) _browser: (NSBrowser*)sender
 selectCellWithString: (NSString*)title
-	inColumn: (int)column
+	inColumn: (NSInteger)column
 {
   NSMatrix *m;
   BOOL isLeaf;
@@ -1605,16 +1605,16 @@ selectCellWithString: (NSString*)title
 @interface NSSavePanel (GSBrowserDelegate)
 - (void) browserDidScroll: (NSBrowser *)sender;
 - (void) browser: (NSBrowser*)sender
-createRowsForColumn: (int)column
+createRowsForColumn: (NSInteger)column
         inMatrix: (NSMatrix*)matrix;
 
 - (BOOL) browser: (NSBrowser*)sender
-   isColumnValid: (int)column;
+   isColumnValid: (NSInteger)column;
 
 - (void) browser: (NSBrowser*)sender
  willDisplayCell: (id)cell
-           atRow: (int)row
-          column: (int)column;
+           atRow: (NSInteger)row
+          column: (NSInteger)column;
 @end 
 
 static int compareFilenames (id elem1, id elem2, void *context)
@@ -1637,7 +1637,7 @@ static int compareFilenames (id elem1, id elem2, void *context)
 }
 
 - (void) browser: (NSBrowser*)sender
-createRowsForColumn: (int)column
+createRowsForColumn: (NSInteger)column
 	inMatrix: (NSMatrix*)matrix
 {
   NSString              *path, *file, *pathAndFile, *extension; 
@@ -1862,7 +1862,7 @@ createRowsForColumn: (int)column
 }
 
 - (BOOL) browser: (NSBrowser*)sender
-   isColumnValid: (int)column
+   isColumnValid: (NSInteger)column
 {
   /*
    * FIXME This code doesn't handle the case where the delegate now wants
@@ -1889,8 +1889,8 @@ createRowsForColumn: (int)column
 
 - (void) browser: (NSBrowser*)sender
  willDisplayCell: (id)cell
-	   atRow: (int)row
-	  column: (int)column
+	   atRow: (NSInteger)row
+	  column: (NSInteger)column
 {
 }
 @end
